backend/requirements.txt
fastapi
uvicorn[standard]
PyPDF2
google-genai
rank-bm25
python-dotenv
python-multipart
